the current time is a double, math on it with a float is a warning.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@16663 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/GuidedMissleStrategy.cxx
@@ -278,7 +278,7 @@ bool GuidedMissileStrategy::_predict(float dt, float p[3], float v[3]) const
   const bool isRemote = (getPath().getPlayer() !=
 			 LocalPlayer::getMyTank()->getId());
 
-  float ctime = currentTime + dt;
+  float ctime = (float)currentTime + dt;
 
   /*
    * If it expires there we'll return false.
